+++ images/base/21.04-ubuntu/CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - cleanup /var/cache/ directory, with fix
 - cleanup /usr/share/doc directory
+- add docker registry branch support